backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
openai>=1.0.0
zhipuai>=2.0.0
python-docx>=0.8.11
reportlab>=4.0.0
playwright>=1.40.0
requests>=2.31.0
